refactor(courses): Follow micro-component architecture
Restructure MembershipPlans and PlanCard to follow strict
micro-component architecture (component-name/index.tsx pattern)
instead of flat files. Update imports accordingly.
