+++ src/app/(frontend)/courses/_components/PlanCard/index.tsx
